pyproject.toml
[project]
name = "robotscan-skill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.1",
]
